Initial refactory - broken
