+++ src/main/frontend/pipeline-console-view/pipeline-console/main/PipelineConsole.tsx
@@ -8,13 +8,6 @@ import {
 } from "../../../pipeline-graph-view/pipeline-graph/main/";
 import { DataTreeView, StepInfo } from "./DataTreeView";
 
-import Typography from "@mui/material/Typography";
-
-import HourglassEmptyIcon from "@mui/icons-material/HourglassEmpty";
-import ScheduleIcon from "@mui/icons-material/Schedule";
-import TimerIcon from "@mui/icons-material/Timer";
-import InfoIcon from "@mui/icons-material/Info";
-import LinkIcon from "@mui/icons-material/Link";
 
 import "./pipeline-console.scss";
 
@@ -29,84 +22,6 @@ interface PipelineConsoleState {
   hasScrolled: boolean;
 }
 
-export interface StageSummaryProps {
-  stage: StageInfo;
-  failedSteps: StepInfo[];
-}
-
-// Tree Item for stages
-const StageSummary = (props: StageSummaryProps) => (
-  <React.Fragment>
-    <div className="stage-detail-group">
-      <Typography color="inherit" className="detail-element-header">
-        Stage '{props.stage.name}'
-      </Typography>
-      <div className="detail-element" key="start-time">
-        <ScheduleIcon className="detail-icon" />
-        {props.stage.startTimeMillis}
-      </div>
-      <div className="detail-element" key="paused-duration">
-        <HourglassEmptyIcon className="detail-icon" />
-        {props.stage.pauseDurationMillis}
-      </div>
-      <div className="detail-element" key="duration">
-        <TimerIcon className="detail-icon" />
-        {props.stage.totalDurationMillis}
-      </div>
-      <div className="detail-element" key="status">
-        <InfoIcon className="detail-icon " />
-        <span className="capitalize">{props.stage.state}</span>
-      </div>
-      {props.failedSteps.map((value: StepInfo) => {
-        console.log(`Found failed step ${value}`);
-        return (
-          <FailedStepLink step={value} key={`failed-step-link-${value.id}`} />
-        );
-      })}
-    </div>
-  </React.Fragment>
-);
-
-export interface StepSummaryProps {
-  step: StepInfo;
-}
-
-// Tree Item for stages
-const StepSummary = (props: StepSummaryProps) => (
-  <React.Fragment>
-    <div className="step-detail-group">
-      <div className="detail-element" key="start-time">
-        <ScheduleIcon className="detail-icon" />
-        {props.step.startTimeMillis}
-      </div>
-      <div className="detail-element" key="paused-duration">
-        <HourglassEmptyIcon className="detail-icon" />
-        {props.step.pauseDurationMillis}
-      </div>
-      <div className="detail-element" key="duration">
-        <TimerIcon className="detail-icon" />
-        {props.step.totalDurationMillis}
-      </div>
-      <div className="detail-element capitalize" key="status">
-        <InfoIcon className="detail-icon" />
-        <span className="capitalize">{props.step.state}</span>
-      </div>
-    </div>
-  </React.Fragment>
-);
-
-export interface FailedStepLinkProps {
-  step: StepInfo;
-}
-
-const FailedStepLink = (props: FailedStepLinkProps) => (
-  <div className="detail-element">
-    <LinkIcon className="detail-icon" />
-    <a className="detail-element" href={`?selected-node=${props.step.id}`}>
-      Failed step: {props.step.name}
-    </a>
-  </div>
-);
 
 export interface ConsoleLineProps {
   lineNumber: string;
@@ -115,28 +30,6 @@ export interface ConsoleLineProps {
   key: string;
 }
 
-// Tree Item for stages
-const ConsoleLine = (props: ConsoleLineProps) => (
-  <div className="console-output-item" key={props.lineNumber}>
-    <div
-      className="console-output-line-anchor"
-      id={`log-${props.lineNumber}`}
-      key={`${props.lineNumber}-anchor`}
-    />
-    <div className="console-output-line" key={`${props.lineNumber}-body`}>
-      <a
-        className="console-line-number"
-        href={`?selected-node=${props.stepId}#log-${props.lineNumber}`}
-      >
-        {props.lineNumber}
-      </a>
-      {makeReactChildren(
-        tokenizeANSIString(props.content),
-        `${props.stepId}-${props.lineNumber}`
-      )}
-    </div>
-  </div>
-);
 
 export class PipelineConsole extends React.Component<
   PipelineConsoleProps,
@@ -302,7 +195,7 @@ export class PipelineConsole extends React.Component<
     return foundStep;
   }
 
-  // Gets the node hierarchy of stages in the tree view (a list of child -> parent -> grandparent).#
+  // Gets the node hierarchy of stages in the tree view (a list of child -> parent -> grandparent).
   // This needs to be given the nodeId of a stage, so call getSelectedStep first to see if the nodeId
   // is a step - and if so pass it step.stageId.
   getStageNodeHierarchy(nodeId: string, stages: StageInfo[]): Array<string> {
@@ -347,70 +240,6 @@ export class PipelineConsole extends React.Component<
     return selectedStep;
   }
 
-  renderStageDetails() {
-    let focusedStage = null;
-    for (let i = 0; i < this.state.stages.length; i++) {
-      let stage = this.state.stages[i];
-      if ("" + stage.id == this.state.selected) {
-        // User has selected a stage node.
-        focusedStage = stage;
-        let failedSteps = [] as StepInfo[];
-        for (let i = 0; i < this.state.steps.length; i++) {
-          let step = this.state.steps[i];
-          if (step.stageId === this.state.selected) {
-            // We seem to get a mix of upper and lower case states, so normalise on lowercase.
-            if (step.state.toLowerCase() === "unstable") {
-              failedSteps.push(step);
-            }
-          }
-        }
-        return (
-          <pre className="console-output">
-            <StageSummary stage={focusedStage} failedSteps={failedSteps} />
-          </pre>
-        );
-      }
-    }
-    return null;
-  }
-
-  renderStepDetails() {
-    for (let i = 0; i < this.state.steps.length; i++) {
-      let step = this.state.steps[i];
-      if ("" + step.id == this.state.selected) {
-        return (
-          <pre className="console-output">
-            <StepSummary step={step} />
-          </pre>
-        );
-      }
-    }
-    return null;
-  }
-
-  renderConsoleOutput() {
-    if (this.state.consoleText.length > 0) {
-      const lineChunks = this.state.consoleText.split("\n");
-      return (
-        <pre className="console-output">
-          {lineChunks.map((line, index) => {
-            let lineNumber = String(index + 1);
-            return (
-              <ConsoleLine
-                content={line}
-                lineNumber={lineNumber}
-                stepId={this.state.selected}
-                key={`${this.state.selected}-${lineNumber}`}
-              />
-            );
-          })}
-        </pre>
-      );
-    } else {
-      return null;
-    }
-  }
-
   render() {
     const splitPaneStyle: React.CSSProperties = {
       position: "relative",
@@ -444,9 +273,7 @@ export class PipelineConsole extends React.Component<
             </div>
 
             <div className="split-pane" key="console-view">
-              {this.renderStageDetails()}
-              {this.renderStepDetails()}
-              {this.renderConsoleOutput()}
+
             </div>
           </SplitPane>
         </div>
